+++ src/main/java/org/swyp/dessertbee/auth/dto/oauth2/OAuth2Response.java
@@ -1,4 +1,4 @@
-package org.swyp.dessertbee.auth.dto;
+package org.swyp.dessertbee.auth.dto.oauth2;
 
 public interface OAuth2Response {
     String getProvider();
