+++ src/main/java/org/swyp/dessertbee/auth/service/KakaoOAuthService.java
@@ -9,8 +9,8 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
-import org.swyp.dessertbee.auth.dto.KakaoResponse;
-import org.swyp.dessertbee.auth.dto.OAuth2Response;
+import org.swyp.dessertbee.auth.dto.oauth2.KakaoResponse;
+import org.swyp.dessertbee.auth.dto.oauth2.OAuth2Response;
 import org.swyp.dessertbee.auth.dto.login.LoginResponse;
 import org.swyp.dessertbee.auth.jwt.JWTUtil;
 import org.swyp.dessertbee.common.entity.ImageType;
